configure hot-module-replacement
