+++ package.json
@@ -7,7 +7,7 @@
   "main": "main.js",
   "scripts": {
     "build": "npm run clean && npm run copy-files && webpack --env.prod",
-    "dev": "webpack-dev-server --env.dev",
+    "dev": "webpack-dev-server --env.dev --hot",
     "e2e-test": "protractor protractor.conf.js",
     "test": "karma start",
     "test-live": "npm test -- --auto-watch --no-single-run",
@@ -86,6 +86,7 @@
     "protractor-cucumber-framework": "=3.1.0",
     "raw-loader": "=0.5.1",
     "react-addons-test-utils": "=15.6.0",
+    "react-hot-loader": "^1.3.1",
     "react-test-renderer": "=15.6.1",
     "should": "=11.2.0",
     "sinon": "=2.0.0",
